crypting onfly affected files and DEBUG

+++ git-encrypt-filter.sh
@@ -2,14 +2,22 @@
 
 IFS="
 "
-echo "recrypt..." >> /tmp/recrypt
-[ -n "$GITCRYPT_ALLFILES" ] && (GITCRYPT_AFFECTED_FILES=`git ls-tree --name-only --full-tree -r $GIT_COMMIT` || exit 1)
-echo "..." >> /tmp/recrypt
 
+_debug() {
+	[ -z "$GITCRYPTDEBUG" ] || echo -e "${1}"
+}
+
+_debug ""
+[ -n "$GITCRYPT_ALLFILES" ] && _debug "  == [filter] ALL files"
+
+if [ -n "$GITCRYPT_ALLFILES" ]; then
+	GITCRYPT_AFFECTED_FILES=`git ls-tree --name-only --full-tree -r $GIT_COMMIT` || exit 1
+fi
 
 for x in $GITCRYPT_AFFECTED_FILES; do
-	echo "recrypt file $x" >> /tmp/recrypt
-	cat "$x" | $@ > "$x.tmp" || continue 
+	cmd="cat \"$x\" | $@ > \"$x.tmp\""
+	_debug "    * $cmd"
+	eval $cmd || continue 
 	cat "$x.tmp" > "$x" 
 	rm -f "$x.tmp"
 done
